feat(gui-client): re-design the "about" window (#9149)
This re-designs the "About" window of the GUI client.

|Before|After|
|---|---|
|![Screenshot From 2025-05-15
11-40-46](https://github.com/user-attachments/assets/ce7f1cf5-e234-4d64-825a-8e1c0badddb5)|![image](https://github.com/user-attachments/assets/a0c2f54f-f8f9-4d57-a8ee-0e826ce004da)|

I also reduced the window in its size and made it non-resizable to avoid
it looking bad when users try to fiddle around with it.

The embedded Git hash is also trimmed to 8 characters, currently it is
embedded as the fully hash.