feat(portal): Add broadcast to simulate more real-world events (#9172)
We are currently consuming the WAL on production and it has shown very
little cost in terms of resource usage.

It would be better to get a more real-world test by sending actual
broadcasts with data.

To do this, we simply send a `Domain.PubSub.broadcast` with all of the
data received in the WAL message, which represents an absolute
worst-case scenario.