refactor(gui-client): update settings window via event (#9193)
Ensuring that the "Settings" window always displays the latest state is
important. At the moment, we achieve this by fetching the settings from
Rust every time we know that they changed and when the window is opened.
Currently, the settings can only change as a result of the reset button.

Once we integrate MDM-controlled configuration into the GUI client, the
settings can change at any point when administrators push a new
configuration change. The window doesn't know that though.

To simplify the code and handle dynamic changes to the settings, we now
use an event to send the settings to the frontend.