fix(android): Fix request permission screen padding (#9200)
The padding on this screen was broken, resulting in elements flush
against the edge of the display.

While I was in here, I decided to do a quick round of polish on the
layout and sizing of elements here.

<img width="395" alt="Screenshot 2025-05-21 at 4 24 08 PM"
src="https://github.com/user-attachments/assets/db4f53a3-1ee9-41ae-8172-8cca22be742b"
/>