fix(ci): Fetch latest PR title for pr-lint (#9225)
When opening a PR with a lengthy title, the `pr-lint` check will fail.
Fixing the title and re-running the check didn't fix it because it was
always referencing the original title that triggered the run.

To fix it, we need to fetch the title fresh on each run of the workflow.