fix(android): ensure connectOnStart applies only to start (#9279)
A bug was introduced in #9227 where the `connectOnStart` was being read
for both resumed launches and initial launches. This PR updates that
logic so that we only check `connectOnStart` when the app is actually
launched.

On mobile platforms, this flag is not as clearcut as on desktop
platforms, but it's maintained here because ChromeOS is also used on
desktop systems.