feat(connlib): expose `isAuthenticationError` over Apple FFI (#9291)
The GUI client already uses the same function to check for
authentication errors. Therefore, this case is already handled there.

For Swift, we can easily expose this getter via the `swift-bridge`
module. For Android, we don't expose it for now. Once we tackle #3959,
this should be easier to do. In the meantime, the UX on Android is not
super terrible. The user gets signed out and we will then receive a 401
when we try to sign-in again the next time.

Resolves: #9289