fix(ci): download musl build deps from google cloud (#9310)
musl.cc wholesale blocked GitHub Actions due to "abuse" so we need a
stop-gap measure to obtain these tools in CI runs.

Unfortunately there are no publicly-listed mirror for these build tools,
and any cron job we attempt to run to update these is going to be
blocked as well, so I simply uploaded them manually to a GCP bucket for
now.